+++ tests/Sniffs/Spacing/SpacingTestSuite.php
@@ -6,11 +6,13 @@ use PHPUnit\Framework\TestSuite;
 
 final class SpacingTestSuite
 {
+
     public function getTestSuite()
     {
         $suite = new TestSuite('XynhaCS_Spacing');
 
         $suite->addTestSuite(PropertySpacingTest::class);
+        $suite->addTestSuite(ClassFunctionSpacingTest::class);
 
         return $suite;
     }